Update: module name from search to media.
