+++ app/build.gradle
@@ -39,7 +39,7 @@ android {
 
 dependencies {
     implementation project(':network')
-    implementation project(':feature:search')
+    implementation project(':feature:media')
     implementation 'androidx.core:core-ktx:1.9.0'
     implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'androidx.activity:activity-ktx:1.6.1'
