+++ feature/media/src/main/AndroidManifest.xml
