+++ feature/media/src/main/kotlin/labs/gas/musical/media/domain/model/DomainErrors.kt
@@ -1,0 +1,7 @@
+package labs.gas.musical.media.domain.model
+
+import labs.gas.musical.network.DomainError
+
+data class SearchDomainError(
+    override val message: String
+) : DomainError
