+++ feature/media/src/main/kotlin/labs/gas/musical/media/presentation/search/SearchFragment.kt
@@ -1,0 +1,24 @@
+package labs.gas.musical.media.presentation.search
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import dagger.hilt.android.AndroidEntryPoint
+import labs.gas.musical.media.R
+
+@AndroidEntryPoint
+class SearchFragment : Fragment() {
+    private val searchViewModel: SearchViewModel by viewModels()
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment_search, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        searchViewModel.search("Yeah yeah yeahs")
+    }
+}
