+++ feature/media/src/main/res/drawable/ic_heart_gray_48dp.xml
