+++ feature/media/src/main/res/drawable/ic_heart_purple_48dp.xml
