+++ feature/media/src/main/res/drawable/ic_launcher_foreground.xml
