+++ feature/media/src/main/res/drawable/ic_media_black_24dp.xml
