+++ feature/media/src/main/res/drawable/ic_search_black_48dp.xml
