+++ feature/media/src/main/res/drawable/shape_search_box.xml
