+++ feature/media/src/main/res/navigation/search_navigation.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/search_navigation"
+    app:startDestination="@id/navigation_fragment">
+
+    <fragment
+        android:id="@+id/navigation_fragment"
+        android:name="labs.gas.musical.media.presentation.search.SearchFragment" />
+
+</navigation>
