--- feature/search/src/main/kotlin/labs/gas/musical/search/domain/model/DomainErrors.kt
@@ -1,7 +1,0 @@
-package labs.gas.musical.search.domain.model
-
-import labs.gas.musical.network.DomainError
-
-data class SearchDomainError(
-    override val message: String
-) : DomainError
