--- feature/search/src/main/kotlin/labs/gas/musical/search/domain/model/DomainModels.kt
@@ -1,11 +1,0 @@
-package labs.gas.musical.search.domain.model
-
-class MediaDomainModel(
-    val id: Int,
-    val photo: String,
-    val name: String?,
-    val album: String?,
-    val artist: String,
-    val duration: Int,
-    var isFavorite: Boolean = false
-)
