--- feature/search/src/main/res/navigation/search_navigation.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/search_navigation"
-    app:startDestination="@id/navigation_fragment">
-
-    <fragment
-        android:id="@+id/navigation_fragment"
-        android:name="labs.gas.musical.search.presentation.SearchFragment" />
-
-</navigation>
