--- feature/search/src/test/kotlin/labs/gas/musical/search/ExampleUnitTest.kt
@@ -1,17 +1,0 @@
-package labs.gas.musical.search
-
-import org.junit.Test
-
-import org.junit.Assert.*
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
