+++ settings.gradle
@@ -15,4 +15,4 @@ dependencyResolutionManagement {
 rootProject.name = "Musical beat"
 include ':app'
 include ':network'
-include ':feature:search'
+include ':feature:media'
